backend/requirements.txt
# Core framework - compatible versions
fastapi>=0.104.0,<0.112.0
uvicorn[standard]>=0.24.0,<0.31.0
httpx>=0.25.0,<0.28.0
python-dotenv>=1.0.0,<2.0.0
cachetools>=5.3.0,<6.0.0

# Database dependencies
sqlalchemy[asyncio]>=2.0.0,<3.0.0
asyncmy>=0.2.8,<0.3.0
alembic>=1.12.0,<2.0.0

# Authentication dependencies
python-multipart>=0.0.6,<1.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-jose[cryptography]>=3.3.0,<4.0.0

# Email dependencies
emails>=0.6.0,<1.0.0
jinja2>=3.1.0,<4.0.0

# Compatibility dependencies
anyio>=3.7.0,<5.0.0
pydantic>=2.5.0,<3.0.0
starlette>=0.27.0,<0.48.0
h11>=0.14.0,<0.17.0

